Moved hashing functions into core. #include<osquery/hash.h>
